+++ docs/CHANGELOG.md
@@ -1,5 +1,5 @@
 ## 0.2.0 (under development)
-- Enh: make icon color configurable
+- Enh: make like icon color configurable
 - Enh: use color and icon pickers
 - Remove default values for configuration and drop support for file configuration
 - Remove verified icon (use https://marketplace.humhub.com/module/verified/description instead)